refactor(portcullis-core): extract IFC lattice to ifc_lattice module (MVK M1)
Move the closure-critical IFC label lattice out of lib.rs into a dedicated
ifc_lattice.rs (552 LOC): ConfLevel, IntegLevel, AuthorityLevel, ProvenanceSet,
Freshness, DerivationClass, IFCLabel + join/flows_to/meet/leq. Re-exported at the
crate root (pub use ifc_lattice::*), so portcullis_core::{IFCLabel, ConfLevel, ...}
resolve exactly as before — zero consumer churn. Added to the kernel-boundary
ratchet's KERNEL_FILES.

This isolates the ~540 verified lattice lines from the other ~3.4k of lib.rs and
brings them under the boundary ratchet (RFC minimum-viable-ifc-kernel.md, M1).
Operation/SinkClass stay in lib.rs — shared with the capability machinery, more
entangled, a later sub-rung.

No behavior change: 754 lib tests pass, all-features build clean, clippy clean,
ratchet green. (The pre-existing provenance_e2e/flow_red_team feature-gated test
compile errors under bare `cargo test` are unrelated and unchanged by this move.)
